requirements.txt
websockets>=12.0
opuslib>=3.0.1
python-dotenv>=1.0
numpy>=1.24
Pillow>=10.0
cairosvg>=2.7
spidev>=3.6
requests>=2.31
paho-mqtt>=1.6
cryptography>=41.0
